requirements.txt
requests
pandas
numpy
tensorflow
scikit-learn
google-cloud-bigquery
google-auth
pandas-gbq
db-dtypes
google-cloud-bigquery-storage
pytest
pylint
ipdb
numpy
pandas
jupyterlab
matplotlib
seaborn
google-cloud-storage
traitlets
db-dtypes
fastapi
uvicorn
joblib
